:construction: manager 객체 index.js에서 분리
