+++ src/index.js
@@ -8,40 +8,7 @@ import Line from "./line.js";
 import { addLineToList, showAllLineInLineManager } from "./line-manager.js";
 import { showLineMenuInSectionManager } from "./section-manager.js";
 import { showMapList } from "./map-print-manager.js";
-
-export function ManageStationLine() {
-  this.lineList = [];
-  this.stationList = [];
-  this.selectedLine = null;
-  this.addStationInList = (station) => {
-    this.stationList.push(station);
-  };
-  this.addLineInList = (line) => {
-    this.lineList.push(line);
-  };
-  this.getAllLineName = () => {
-    let lineNameList = [];
-    for (let i in lineList) {
-      lineNameList.push(this.lineList[i].name);
-    }
-
-    return lineNameList;
-  };
-  this.setSelectedLine = (selectedLine) => {
-    this.selectedLine = selectedLine;
-  };
-  this.getSelectedLine = () => {
-    return this.selectedLine;
-  };
-  this.setChangedLine = (changedLine) => {
-    this.lineList.forEach((line) => {
-      if (line.name === changedLine.name) {
-        line = changedLine;
-      }
-    });
-  };
-}
-export const manager = new ManageStationLine();
+import { manager } from "./manager.js";
 
 const btnStationManager = document.getElementById("station-manager-button");
 const btnLineManager = document.getElementById("line-manager-button");
@@ -104,7 +71,7 @@ let stationList = [
   new Station("강남"),
 ];
 for (let i in stationList) {
-  manager.addStationInList(stationList[i]);
+  manager.setStationInManager(stationList[i]);
 }
 
 btnAddStation.onclick = () => {
@@ -119,13 +86,13 @@ btnAddStation.onclick = () => {
 
 const tmp = new Line("2호선");
 tmp.addLine(new Station("잠실"), new Station("신림"));
-manager.addLineInList(tmp);
+manager.setLineInManager(tmp);
 const tmp2 = new Line("3호선");
 tmp2.addLine(new Station("오금"), new Station("대화"));
-manager.addLineInList(tmp2);
+manager.setLineInManager(tmp2);
 const tmp3 = new Line("4호선");
 tmp3.addLine(new Station("당고개"), new Station("오이도"));
-manager.addLineInList(tmp3);
+manager.setLineInManager(tmp3);
 
 const btnAddLine = document.getElementById("line-add-button");
 btnAddLine.onclick = () => {
